counts fitting function

+++ digamma_pk.m
@@ -1,4 +1,4 @@
 function Y = digamma_pk(X,mu,s)
 
-Q = sum(exp(-mu*digamma((min(X)>0):max(X),s)));
+Q = sum(exp(-mu*digamma(double(min(X)>0):max(X),s)));
 Y = exp(-mu*digamma(X,s))/Q;
